Removed unused import, added all schemas to the global file
