+++ noq_django/rest_api/api/admin_activities_api.py
@@ -1,56 +1,22 @@
-from datetime import datetime, date
-from typing import Dict, List
-from ninja import Router, ModelSchema, Schema
-from backend.models import Activity, VolunteerActivity, VolunteerProfile
+from typing import List
+from ninja import Router
+from backend.models import Activity, VolunteerActivity
 from django.shortcuts import get_object_or_404
 
 from .api_schemas import ActivityUpdateSchema
 
+
+from .api_schemas import (
+    ActivitySchema,
+    ActivityDetailSchema,
+    ActivityCreateSchema,
+    ActivityUpdateSchema,
+    ActivityListSchema
+)
+
+#TODO This needs to be changed !!!!!!!!
 router = Router()
 
-#---- API SCHEMAS ----#
-class ActivitySchema(ModelSchema):
-    class Config:
-        model = Activity
-        model_fields = ["id", "title", "description", "start_time", "end_time", "is_approved"]
-
-class VolunteerProfileSchema(Schema):
-    id: int
-    first_name: str
-    last_name: str
-    email: str
-    date_joined: date
-    registered_at: date
-
-class ActivityDetailSchema(ActivitySchema):
-    volunteers: List[VolunteerProfileSchema]
-
-class ActivityCreateSchema(ModelSchema):
-    class Config:
-        model = Activity
-        model_fields = ["title", "description", "start_time", "end_time", "is_approved"]
-
-class ActivityUpdateSchema(ModelSchema):
-    class Config:
-        model = Activity
-        model_fields = ["title", "description", "start_time", "end_time", "is_approved"]
-
-
-class SimpleVolunteerSchema(Schema):
-    id: int
-    first_name: str
-    last_name: str
-
-class ActivityListSchema(Schema):
-    id: int
-    title: str
-    description: str
-    start_time: datetime
-    end_time: datetime
-    is_approved: bool
-    volunteers: List[SimpleVolunteerSchema]
-
-#---- API ENDPOINTS ----#
 
 # List all activities without volunteers' detail
 @router.get("/", response=List[ActivityListSchema], tags=["Admin Activities"])
